Silence warnings.
[git-p4: depot-paths = "//depot/qt/main/": change = 94850]

+++ qmake/project.cpp
@@ -109,7 +109,7 @@ static QStringList split_value_list(const QString &vals, bool do_semicolon=FALSE
     QStringList ret;
     QValueStack<QChar> quote;
     for(int x = 0; x < (int)vals.length(); x++) {
-	if(x != vals.length()-1 && vals[x] == '\\' && (vals[x+1] == '\'' || vals[x+1] == '"'))
+	if(x != (int)vals.length()-1 && vals[x] == '\\' && (vals[x+1] == '\'' || vals[x+1] == '"'))
 	    build += vals[x++]; //get that 'escape'
 	else if(!quote.isEmpty() && vals[x] == quote.top())
 	    quote.pop();
